Streamlit/requirements.txt
streamlit
pandas
plotly
sqlalchemy
psycopg2-binary
python-dotenv
python-dateutil
numpy
